오류수정

+++ src/main/java/com/dksys/biz/user/ar/ar01/service/impl/AR01SvcImpl.java
@@ -232,7 +232,7 @@ public class AR01SvcImpl implements AR01Svc {
 			paramMap.put("trstRprcSeq", detailMap.get("shipSeq"));		
 			paramMap.put("trstDtlSeq", detailMap.get("shipDtlSeq"));				  	
 			paramMap.put("odrNo", paramMap.get("odrSeq"));
-			int bilgVatAmt = ar02Mapper.selectBilgVatAmt(paramMap);
+			long bilgVatAmt = ar02Mapper.selectBilgVatAmt(paramMap);
 			paramMap.put("bilgVatAmt", String.valueOf(bilgVatAmt));
 			realTotTrstAmt += bilgVatAmt;
 			ar02Mapper.insertPchsSell(paramMap);
